+++ ContentPatcher/release-notes.md
@@ -7,6 +7,7 @@
   * `SkillLevel`: the player's level for a given skill.
 * Added `Wind` value for `Weather` token.
 * Added support for matching subkey/value pairs for tokens which previously required subkeys.
+* Improved performance for most content packs.
 * Fixed language token always marked 'not valid in this context'.
 * Fixed token strings not validated for format version compatibility.
 * Fixed some 1.5 tokens not validated for format version compatibility.